+++ api/src/lib/format-error.js
@@ -1,8 +1,9 @@
 var Boom = require('boom');
 
 module.exports = function(error) {
-  if (! error.data.details) {
-    return Boom.badImplementation(error);
+  if (!error.data || !error.data.details) {
+    console.error(error);
+    return error.isBoom ? error : Boom.badImplementation(error);
   }
 
   error.output.payload.validationErrors = error.data.details.map(failure => ({