+++ api/src/plugins/openid-connect/grants/password.js
@@ -1,34 +1,43 @@
 module.exports = (options) => ({
   params: ['username', 'password'],
   grantTypeFactory: function passwordGrantTypeFactory(providerInstance) {
-    return function * passwordGrantType(next) {
-      const { username, password } = this.oidc.params;
-      const account = yield options.authenticateUser(username, password);
+    return async function passwordGrantType(ctx, next) {
+      const { username, password } = ctx.oidc.params;
+      const account = await options.authenticateUser(username, password);
 
       if (account) {
-        const AccessToken = providerInstance.AccessToken;
+        const { AccessToken, IdToken } = providerInstance;
         const at = new AccessToken({
           accountId: 'foo',
-          clientId: this.oidc.client.clientId,
-          grantId: this.oidc.uuid,
+          clientId: ctx.oidc.client.clientId,
+          grantId: ctx.oidc.uuid,
         });
 
-        const accessToken = yield at.save();
+        const accessToken = await at.save();
         const expiresIn = AccessToken.expiresIn;
 
-        this.body = {
+        const token = new IdToken(
+          Object.assign({}, await Promise.resolve(account.claims())),
+          ctx.oidc.client.sectorIdentifier
+        );
+        token.set('at_hash', accessToken);
+
+        const idToken = await token.sign(ctx.oidc.client);
+
+        ctx.body = {
           access_token: accessToken,
           expires_in: expiresIn,
           token_type: 'Bearer',
+          id_token: idToken,
         };
       } else {
-        this.body = {
+        ctx.body = {
           error: 'invalid_grant',
           error_description: 'invalid credentials provided',
         };
       }
 
-      yield next;
+      await next();
     };
   }
 });