+++ api/src/plugins/openid-connect/openid-connect.js
@@ -1,86 +1,19 @@
 'use strict';
 const OidcProvider = require('oidc-provider');
-const cors = require('koa-cors');
-const querystring = require('querystring');
-const path = require('path');
-const fs = require('fs');
-const handlebars = require('handlebars');
+const cors = require('koa2-cors');
 const passwordGrant = require('./grants/password');
+const getConfig = require('./getConfig');
+const addRoutes = require('./addRoutes');
 
 exports.register = function (server, options, next) {
   const issuer = options.issuer || process.env.OIDC_BASE_URL || 'http://localhost:9000';
 
   const prefix = options.prefix ? `/${options.prefix}` : '/op';
-  const provider = new OidcProvider(issuer, {
-    adapter: options.adapter,
-    findById: options.findUserById,
-    routes: {
-      authorization: `${prefix}/auth`,
-      certificates: `${prefix}/certs`,
-      check_session: `${prefix}/session/check`,
-      end_session: `${prefix}/session/end`,
-      introspection: `${prefix}/token/introspection`,
-      registration: `${prefix}/reg`,
-      revocation: `${prefix}/token/revocation`,
-      token: `${prefix}/token`,
-      userinfo: `${prefix}/me`,
-    },
-    acrValues: ['session', 'urn:mace:incommon:iap:bronze'],
-    cookies: {
-      long: { signed: true },
-      short: { signed: true },
-    },
-    discovery: {
-      service_documentation: '',
-      version: '',
-    },
-    claims: {
-      amr: null,
-      address: ['address'],
-      email: ['email', 'email_verified'],
-      phone: ['phone_number', 'phone_number_verified'],
-      app_metadata: ['app_metadata'],
-      profile: ['birthdate', 'family_name', 'gender', 'name', 'given_name', 'locale', 'middle_name', 'name',
-        'nickname', 'picture', 'preferred_username', 'profile', 'updated_at', 'website', 'zoneinfo'],
-    },
-    features: {
-      devInteractions: false,
-      discovery: true,
-      claimsParameter: true,
-      clientCredentials: true,
-      encryption: true,
-      introspection: true,
-      oauthNativeApps: true,
-      registration: {
-        initialAccessToken: options.initialAccessToken,
-      },
-      registrationManagement: false,
-      request: true,
-      requestUri: true,
-      revocation: true,
-      sessionManagement: true,
-      backchannelLogout: false,
-    },
-    logoutSource: function renderLogoutSource(form) {
-      // this => koa context;
-      const layout = fs.readFileSync(path.join(__dirname, './templates/layout/layout.hbs'), 'utf8');
-      const logout = fs.readFileSync(path.join(__dirname, './templates/end_session.hbs'), 'utf8');
-
-      this.body = handlebars.compile(layout)({
-        content: handlebars.compile(logout)({ form })
-      });
-    },
-    subjectTypes: ['public', 'pairwise'],
-    pairwiseSalt: 'da1c442b365b563dfc121f285a11eedee5bbff7110d55c88',
-    interactionUrl: function interactionUrl(interaction) {
-      // this => koa context;
-      return `/interaction/${this.oidc.uuid}`;
-    },
-  });
+  const provider = new OidcProvider(issuer, getConfig(options));
 
   provider.initialize({
-    keystore: options.keystores.certificates,
-    integrity: options.keystores.integrity,
+    adapter: options.adapter,
+    keystore: options.keystore,
   })
     .then(() => {
       const { grantTypeFactory, params } = passwordGrant(options);
@@ -97,113 +30,8 @@ exports.register = function (server, options, next) {
         }
       });
 
-      server.views({
-        engines: {
-          hbs: handlebars
-        },
-        relativeTo: __dirname,
-        path: 'templates',
-        layout: true,
-        layoutPath: './templates/layout',
-      });
-
-      server.route({
-        method: 'GET',
-        path: '/interaction/{grant}',
-        handler: (request, reply) => {
-          const cookie = provider.interactionDetails(request.raw.req);
-          const client = provider.Client.find(cookie.params.client_id);
-
-          if (cookie.interaction.error === 'login_required') {
-            reply.view('login', {
-              client,
-              cookie,
-              title: 'Log In',
-              debug: querystring.stringify(cookie.params, ',<br/>', ' = ', {
-                encodeURIComponent: value => value,
-              }),
-              interaction: querystring.stringify(cookie.interaction, ',<br/>', ' = ', {
-                encodeURIComponent: value => value,
-              }),
-              forgotPasswordPath: `${issuer}/user/forgot-password?client_id=${cookie.params.client_id}&response_type=${cookie.params.response_type}&scope=${cookie.params.scope}&redirect_uri=${cookie.params.redirect_uri}&nonce=${cookie.params.nonce}`,
-            });
-          } else {
-            reply.view('interaction', {
-              client,
-              cookie,
-              title: 'Authorize',
-              debug: querystring.stringify(cookie.params, ',<br />', ' = ', {
-                encodeURIComponent: value => value,
-              }),
-              interaction: querystring.stringify(cookie.interaction, ',<br/>', ' = ', {
-                encodeURIComponent: value => value,
-              }),
-            });
-          }
-        },
-        config: {
-          state: {
-            parse: false // hapi fails to parse oidc cookie...
-          }
-        }
-      });
-
-      server.route({
-        method: 'POST',
-        path: '/interaction/{grant}/confirm',
-        handler: (request, reply) => {
-          const result = { consent: {} };
-          provider.interactionFinished(request.raw.req, request.raw.res, result);
-        },
-        config: {
-          state: {
-            parse: false
-          }
-        }
-      });
-
-      server.route({
-        method: 'POST',
-        path: '/interaction/{grant}/login',
-        handler: (request, reply) => {
-          options.authenticateUser(request.payload.login, request.payload.password)
-            .then(account => {
-              if (account) {
-                const result = {
-                  login: {
-                    account: account.accountId,
-                    acr: 'urn:mace:incommon:iap:bronze',
-                    amr: ['pwd'],
-                    remember: request.payload.remember,
-                    ts: Math.floor(Date.now() / 1000),
-                  },
-                  consent: {}
-                };
-                provider.interactionFinished(request.raw.req, request.raw.res, result);
-              } else {
-                const cookie = provider.interactionDetails(request.raw.req);
-                const client = provider.Client.find(cookie.params.client_id);
-                reply.view('login', {
-                  error: 'Invalid email password combination',
-                  client,
-                  cookie,
-                  title: 'Log In',
-                  debug: querystring.stringify(cookie.params, ',<br/>', ' = ', {
-                    encodeURIComponent: value => value,
-                  }),
-                  interaction: querystring.stringify(cookie.interaction, ',<br/>', ' = ', {
-                    encodeURIComponent: value => value,
-                  }),
-                });
-              }
-            });
-        },
-        config: {
-          state: {
-            parse: false
-          }
-        }
-      });
+      server.views(options.vision);
+      addRoutes(server, issuer, options)
 
       server.expose('provider', provider);