+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-## [v1.4.0](https://github.com/synapsestudios/oidc-platform/compare/v1.3.1...release/v1.4.0) - 2017-10-3
+## [v2.0.0](https://github.com/synapsestudios/oidc-platform/compare/v1.4.0...v2.0.0) - 2017-10-10
+### Changed
+- [#162](https://github.com/synapsestudios/oidc-platform/issues/162) Stop allowing the access token for auth on the profile screen
+- [#138](https://github.com/synapsestudios/oidc-platform/issues/138) Stop hardcoding nonce when resetting passwords
+- [#119](https://github.com/synapsestudios/oidc-platform/issues/119) Updated to to [node-oidc](https://github.com/panva/node-oidc-provider) v2.3.2
+### Added
+- [#115](https://github.com/synapsestudios/oidc-platform/issues/115) Theming. See [theming guide](docs/theming.md) for details
+- [#131](https://github.com/synapsestudios/oidc-platform/issues/131) Password Grant now returns an id token as well as an access token
+
+## [v1.4.0](https://github.com/synapsestudios/oidc-platform/compare/v1.3.1...v1.4.0) - 2017-10-3
 ### Added
 - [#127](https://github.com/synapsestudios/oidc-platform/pull/127) Update mysql config to be closer to AWS RDS mysql
 
-## [v1.3.1](https://github.com/synapsestudios/oidc-platform/compare/v1.3.0...release/v1.3.1) - 2017-10-2
+## [v1.3.1](https://github.com/synapsestudios/oidc-platform/compare/v1.3.0...v1.3.1) - 2017-10-2
 ### Added
 - [#123](https://github.com/synapsestudios/oidc-platform/pull/123) Allow password grant type in database.